api/requirements.txt
basf_auth
basf-opti>=0.8.5
basf-pyreto>=0.2.2
botorch==0.3.3
celery<5.0.0
click==7.*
entmoot>=0.2.2
fastapi
gpytorch==1.3.0
gurobipy>=9.1.2
hvwfg
mongomock
numpy
orjson
pandas
plotly>=4.13
pymongo
pymoo
pyrff
redis
scikit-learn
torch<1.9
tqdm